Refactor rename things
